+++ CHANGELOG.md
@@ -19,6 +19,10 @@ Current
 
 #### Changed:
 
+-  [Fix Dimension Serialization Problem with Nested Queries](https://github.com/yahoo/fili/pull/15)
+    * Modified `DimensionToDefaultDimensionSpec` serializer to serialize dimension to apiName if it is not the inner most query
+    * Added helper `hasInnerQuery` to `Util` in serializer package to determine if current query is the inner most query or not
+    * Added tests for `DimensionToDefaultDimensionSpec`
 
 ### Fixed: